pyproject.toml
[project]
name = "stella"
version = "2.0.0"
description = "Stella - WhatsApp Sales Agent for Strides"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "motor>=3.6",
    "pydantic-settings>=2.6",
    "httpx>=0.28",
    "openai>=1.60",
    "anthropic>=0.42",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mongomock-motor>=0.0.34",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"app/llm/prompts/*.py" = ["E501"]
"app/fsm/handlers/price_fallback.py" = ["E501"]
"app/fsm/handlers/closing.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
